Refactoring ★
