+++ src/ignitus-UserInterfaceKit/Components/Organisms/interfaceCardLayout/Components/interfaceCardLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {StyledHeading2} from '../../../userInterfaceKit/styles';
+import {StyledHeading2} from '../../../../styles';
 import {CardLayoutContainer} from '../style';
 import Opportunity from '../../../../../ignitus-Shared/ignitus-DesignSystem/ignitus-Organisms/ignitus-OppurtunityList/Components/OppurtunityList';
