+++ src/ignitus-UserInterfaceKit/InterfaceSideNavigation/constants.ts
