+++ src/ignitus-UserInterfaceKit/InterfaceSideNavigation/style.ts
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled';
-import {Heading3} from '../../../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/atoms';
-import * as C from '../../../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/colors';
-import Icon from '../../../ignitus-Shared/ignitus-Utilities/Components/icon';
-import {flexibleRowDiv} from '../../../ignitus-Shared/ignitus-DesignSystem/shared';
+import {Heading3} from '../../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/atoms';
+import * as C from '../../ignitus-Shared/ignitus-DesignSystem/ignitus-Atoms/colors';
+import Icon from '../../ignitus-Shared/ignitus-Utilities/Components/icon';
+import {flexibleRowDiv} from '../../ignitus-Shared/ignitus-DesignSystem/shared';
 
 type ArrowProps = {
   isexpanded: boolean;
